# [perf] Refactor `mkInferno` to share evaluation of prelude