--- a/inferno-core/app/Main.hs
+++ b/inferno-core/app/Main.hs
@@ -2,9 +2,9 @@
 
 module Main where
 
-import qualified Data.Map as M
+import qualified Data.Map as Map
 import qualified Data.Text.IO as Text
-import Inferno.Core (Interpreter (evalInEnv, parseAndInferTypeReps), mkInferno)
+import Inferno.Core (Interpreter (evalExpr, parseAndInferTypeReps), defaultEnv, mkInferno)
 import Inferno.Module.Prelude (builtinModules)
 import Inferno.Utils.Prettyprinter (showPretty)
 import System.Environment (getArgs)
@@ -13,10 +13,10 @@ main :: IO ()
 main = do
   file <- head <$> getArgs
   src <- Text.readFile file
-  let inferno = mkInferno builtinModules :: Interpreter ()
+  inferno <- mkInferno builtinModules :: IO (Interpreter ())
   case parseAndInferTypeReps inferno src of
     Left err -> print err
     Right ast ->
-      evalInEnv inferno M.empty M.empty ast >>= \case
+      evalExpr inferno (defaultEnv inferno) Map.empty ast >>= \case
         Left err -> print err
         Right res -> showPretty res