--- a/inferno-core/app/Main.hs
+++ b/inferno-core/app/Main.hs
@@ -1,10 +1,11 @@
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NamedFieldPuns #-}
 
 module Main where
 
 import qualified Data.Map as Map
 import qualified Data.Text.IO as Text
-import Inferno.Core (Interpreter (evalExpr, parseAndInferTypeReps), defaultEnv, mkInferno)
+import Inferno.Core (Interpreter (..), mkInferno)
 import Inferno.Module.Prelude (builtinModules)
 import Inferno.Utils.Prettyprinter (showPretty)
 import System.Environment (getArgs)
@@ -15,13 +16,14 @@ main :: IO ()
 main = do
   file <- head <$> getArgs
   src <- Text.readFile file
-  inferno <- mkInferno builtinModules :: IO (Interpreter ())
-  case parseAndInferTypeReps inferno src of
+  Interpreter {evalExpr, defaultEnv, parseAndInferTypeReps} <-
+    mkInferno builtinModules :: IO (Interpreter ())
+  case parseAndInferTypeReps src of
     Left err -> do
       hPutStrLn stderr $ show err
       exitFailure
     Right ast ->
-      evalExpr inferno (defaultEnv inferno) Map.empty ast >>= \case
+      evalExpr defaultEnv Map.empty ast >>= \case
         Left err -> do
           hPutStrLn stderr $ show err
           exitFailure