--- a/inferno-core/src/Inferno/Core.hs
+++ b/inferno-core/src/Inferno/Core.hs
@@ -35,21 +35,10 @@ data InfernoError
 
 -- | Public API for the Inferno interpreter. @c@ is the custom value type.
 data Interpreter c = Interpreter
-  { -- | Evaluates an Expr in a given (local) pinned env and implicit env
-    evalInEnv ::
-      forall a.
-      Map.Map ExtIdent (Value c (ImplEnvM IO c)) ->
-      Map.Map ExtIdent (Value c (ImplEnvM IO c)) ->
-      Expr (Maybe VCObjectHash) a ->
-      IO (Either EvalError (Value c (ImplEnvM IO c))),
-    -- | Evaluates an Expr in an empty pinned and implicit env. This will be
-    -- much faster than @evalInEnv Map.empty Map.empty@ as it reuses the
-    -- evaluation of the prelude.
-    evalInEmptyEnv ::
-      forall a.
-      Expr (Maybe VCObjectHash) a ->
-      IO (Either EvalError (Value c (ImplEnvM IO c))),
-    evalInImplEnvM ::
+  { -- | Evaluates an Expr in a given pinned and implicit env. Use
+    -- @defaultEnv@ for an empty env (only prelude) or compute one using
+    -- @mkEnvFromClosure@.
+    evalExpr ::
       forall a.
       TermEnv VCObjectHash c (ImplEnvM IO c) ->
       Map.Map ExtIdent (Value c (ImplEnvM IO c)) ->
@@ -62,26 +51,31 @@ data Interpreter c = Interpreter
       Text ->
       Either InfernoError (Expr (Pinned VCObjectHash) SourcePos, TCScheme),
     -- | Evaluates all functions in given closure and creates a pinned env containing them
-    mkPinnedEnvFromClosure ::
+    mkEnvFromClosure ::
       Map.Map ExtIdent (Value c (ImplEnvM IO c)) ->
       Map.Map VCObjectHash VCObject ->
-      ImplEnvM IO c (TermEnv VCObjectHash c (ImplEnvM IO c))
+      ImplEnvM IO c (TermEnv VCObjectHash c (ImplEnvM IO c)),
+    -- | The default pinned env containing only the prelude
+    defaultEnv ::
+      TermEnv VCObjectHash c (ImplEnvM IO c)
   }
 
 mkInferno :: forall c. (Eq c, Pretty c) => ModuleMap IO c -> IO (Interpreter c)
 mkInferno prelude = do
-  emptyTermEnv <- runImplEnvM Map.empty $ builtinModulesTerms prelude
+  -- We pre-compute envs that only depend on the prelude so that they can be
+  -- shared among evaluations of different scripts
+  (preludeIdentEnv, preludePinnedEnv) <- runImplEnvM Map.empty $ builtinModulesTerms prelude
   return $
     Interpreter
-      { evalInEnv =
-          \localEnv env expr ->
-            mkTermEnv localEnv
-              >>= \lenv -> runEvalIO lenv env expr,
-        evalInEmptyEnv = runEvalIO emptyTermEnv Map.empty,
-        evalInImplEnvM = runEvalIO,
+      { -- { evalInEnv =
+        --     \localEnv env expr ->
+        --       mkTermEnv localEnv
+        --         >>= \lenv -> runEvalIO lenv env expr,
+        evalExpr = runEvalIO,
         parseAndInferTypeReps = parseAndInferTypeReps,
         parseAndInfer = parseAndInfer,
-        mkPinnedEnvFromClosure = mkPinnedEnvFromClosure
+        mkEnvFromClosure = mkEnvFromClosure preludePinnedEnv,
+        defaultEnv = (preludeIdentEnv, preludePinnedEnv)
       }
   where
     parseAndInfer src =
@@ -119,13 +113,11 @@ mkInferno prelude = do
 
     allClasses = Set.unions $ moduleTypeClasses builtinModule : [cls | Module {moduleTypeClasses = cls} <- Map.elems prelude]
 
-    mkTermEnv localEnv = runImplEnvM localEnv $ builtinModulesTerms prelude
+    -- mkTermEnv localEnv = runImplEnvM localEnv $ builtinModulesTerms prelude
 
     -- TODO at some point: instead of evaluating closure and putting into pinned env,
     -- add closure into the expression being evaluated by using let bindings.
-    -- Also, assert that `fst <$> builtinModulesTerms prelude` is empty before discarding
-    mkPinnedEnvFromClosure localEnv closure = do
-      pinnedEnv <- snd <$> builtinModulesTerms prelude
+    mkEnvFromClosure preludePinnedEnv localEnv closure = do
       mdo
         pinnedEnv' <-
           foldM
@@ -138,6 +130,6 @@ mkInferno prelude = do
                       pure $ Map.insert hash val env
                 _ -> pure env
             )
-            pinnedEnv
+            preludePinnedEnv
             (Map.toList closure)
         pure (localEnv, pinnedEnv')