--- a/inferno-lsp/inferno-lsp.cabal
+++ b/inferno-lsp/inferno-lsp.cabal
@@ -1,6 +1,6 @@
 cabal-version:       >=1.10
 name:                inferno-lsp
-version:             0.1.9
+version:             0.1.10
 synopsis:            LSP for Inferno
 description:         A language server protocol implementation for the Inferno language
 category:            IDE,DSL,Scripting
@@ -33,7 +33,7 @@ library
     , co-log-core                        >= 0.3.1 && < 0.4
     , containers                         >= 0.6.2 && < 0.7
     , exceptions                         >= 0.10.4 && < 0.11
-    , inferno-core                       >= 0.3.0 && < 0.5
+    , inferno-core                       >= 0.3.0 && < 0.6
     , inferno-types                      >= 0.2.0 && < 0.3
     , inferno-vc                         >= 0.3.0 && < 0.4
     , lsp                                >= 1.6.0 && < 1.7