--- a/inferno-ml-remote/inferno-ml-remote.cabal
+++ b/inferno-ml-remote/inferno-ml-remote.cabal
@@ -1,6 +1,6 @@
 cabal-version:      2.4
 name:               inferno-ml-remote
-version:            0.1.0
+version:            0.1.1
 synopsis:           Server for Inferno ML
 description:        Server for Inferno ML
 homepage:           https://github.com/plow-technologies/inferno.git#readme