--- a/inferno-ml/app/Main.hs
+++ b/inferno-ml/app/Main.hs
@@ -2,9 +2,9 @@
 
 module Main where
 
-import qualified Data.Map as M
+import qualified Data.Map as Map
 import qualified Data.Text.IO as Text
-import Inferno.Core (Interpreter (evalInEnv, parseAndInferTypeReps), mkInferno)
+import Inferno.Core (Interpreter (defaultEnv, evalExpr, parseAndInferTypeReps), mkInferno)
 import Inferno.ML.Module.Prelude (mlPrelude)
 import Inferno.ML.Types.Value (MlValue)
 import Inferno.Utils.Prettyprinter (showPretty)
@@ -14,10 +14,10 @@ main :: IO ()
 main = do
   file <- head <$> getArgs
   src <- Text.readFile file
-  let inferno = mkInferno mlPrelude :: Interpreter MlValue
+  inferno <- mkInferno mlPrelude :: IO (Interpreter MlValue)
   case parseAndInferTypeReps inferno src of
     Left err -> print err
     Right ast ->
-      evalInEnv inferno M.empty M.empty ast >>= \case
+      evalExpr inferno (defaultEnv inferno) Map.empty ast >>= \case
         Left err -> print err
         Right res -> showPretty res